# Add metatables to lua vectors