--- a/builtin/mainmenu/tests/serverlistmgr_spec.lua
+++ b/builtin/mainmenu/tests/serverlistmgr_spec.lua
@@ -2,6 +2,7 @@ _G.core = {}
 _G.unpack = table.unpack
 _G.serverlistmgr = {}
 
+dofile("builtin/common/vector.lua")
 dofile("builtin/common/misc_helpers.lua")
 dofile("builtin/mainmenu/serverlistmgr.lua")