- This PR is basically the same as #10330.
- Goal of the PR: Add metatables to vectors, so you can more easily program with them.
- Included features:
  - All the builtin helpers and also minetest callbacks (like `can_dig`) give vectors with metatables.
  - `:`-syntax: `local a = b:offset(0,1,0)`
  - operators: `local a = -b * 2`
  - indexing with numbers: `local a = v[1]`
  - Many unittests.
- backwards compatibility:
  - There are no weird hacks for keeping backwards compatibility.
  - If you want to use `==`, `:` or the number-indexing on vectors that you've got from other mods, you will probably have to copy them via `vector.new` first.
  - Old code should still work fine, with very few exceptions, eg.: mods that store something in indices `1`..`3` in vectors generated by vector helpers, and mods that depend on vectors generated by vector helpers to have no metatable (something like `assert(getmetatable(v)==nil)`. I highly doubt that such mods are common.
- usecases? Copied from #9612:

old code | new code
----------- | ------
`vector.multiply(vector.add(vector.multiply(a, -1), b), 2)` | `(-a + b) * 2`
`vector.length(a)` | `a:length()`
`local xyz = {"x", "y", "z"} for i = 1, 3 do if v[xyz[i]] == 0 then print("bla") end end` | `for i = 1, 3 do if v[i] == 0 then print("bla") end end`
`if vector.equal(a, b) then` | `if a == b then`
`print(vector.to_string(v))` | `print(v)`

### Performance

TL;DR:
- Imo it's ok. :green_square: 
- At least in luajit if you do stuff in a long loop, the performance of the old operations is not worse.
- The number-indexing is slower.
- In certain situations the new stuff can be faster.

<details>

I've compared master with this PR, using luajit2.1, CPU was clocked at 1600MHz. 

#### code:
```lua

local insec_env = minetest.request_insecure_environment()
--~ local jdump = insec_env.require("jit.dump")
local jdump = insec_env.require("jit.v")

--~ local is_new = true
local is_new = false

local function meassure(name, func, new_stuff_needed, ...)
	if new_stuff_needed and not is_new then
		return
	end
	print(name..":")
	local time_start = minetest.get_us_time()
	--~ jdump.on()
	func(...)
	--~ jdump.off()
	local time_end = minetest.get_us_time()
	print("time taken: " .. ((time_end - time_start) / 1000) .. " ms\n")
end

local is_luajit = rawget(_G, "jit") ~= nil

print("benchmarking...")
print("Lua: " .. (is_luajit and jit.version or "PUC Lua"))
print("")

local n1 = is_luajit and 2000000 or 500000
if false then
	meassure("vec_new_i1", function()
		local v1s = {}
		local v2s = {}
		local v3s = {}
		for i = 1, n1 do
			local v = vector.new(i, i, i)
			--~ local v = vector.from(i, i, i)
			v1s[i] = v
			--~ v2s[i] = vector.copy(v)
			--~ v3s[i] = vector.zero()
			v2s[i] = vector.new(v)
			v3s[i] = vector.new()
		end
		print(#v1s)
		print(#v2s)
		print(#v3s)
	end, true)
end

local n3 = is_luajit and 1000000 or 5000000
local rand_vecs = {}
meassure("create_rand_vecs", function()
	for i = 0, n3 do
		rand_vecs[i] = vector.new(math.random(), math.random(), math.random())
	end
end)

local n2 = is_luajit and 100000000 or 5000000
meassure("vec_add_loop1_.add", function()
	local sum = vector.new()
	for i = 1, n2 do
		sum = vector.add(sum, rand_vecs[i%n3])
	end
	print(sum.x)
end)
meassure("vec_add_loop2_:add", function()
	local sum = vector.new()
	for i = 1, n2 do
		sum = sum:add(rand_vecs[i%n3])
	end
	print(sum.x)
end, true)
meassure("vec_add_loop3_+", function()
	local sum = vector.new()
	for i = 1, n2 do
		sum = sum + rand_vecs[i%n3]
	end
	print(sum.x)
end, true)
meassure("vec_add_loop1_.add", function()
	local sum = vector.new()
	for i = 1, n2 do
		sum = vector.add(sum, rand_vecs[i%n3])
	end
	print(sum.x)
end)
meassure("vec_add_loop1_.add", function()
	local sum = vector.new()
	for i = 1, n2 do
		sum = vector.add(sum, rand_vecs[i%n3])
	end
	print(sum.x)
end)

local n21 = is_luajit and 10000000 or 500000
meassure("vec_add2_loop1_.add", function()
	local sum = vector.new()
	for i = 1, n21 do
		local k = (i+1)%n3
		rand_vecs[k] = vector.add(rand_vecs[k], rand_vecs[i%n3])
	end
	print(sum.x)
end)
meassure("vec_add2_loop2_:add", function()
	local sum = vector.new()
	for i = 1, n21 do
		local k = (i+1)%n3
		rand_vecs[k] = rand_vecs[k]:add(rand_vecs[i%n3])
	end
	print(sum.x)
end, true)
meassure("vec_add2_loop3_+", function()
	local sum = vector.new()
	for i = 1, n21 do
		local k = (i+1)%n3
		rand_vecs[k] = rand_vecs[k] + rand_vecs[i%n3]
	end
	print(sum.x)
end, true)
meassure("vec_add2_loop1_.add", function()
	local sum = vector.new()
	for i = 1, n21 do
		local k = (i+1)%n3
		rand_vecs[k] = vector.add(rand_vecs[k], rand_vecs[i%n3])
	end
	print(sum.x)
end)

local n4 = is_luajit and 100000000 or 5000000
meassure("vec_index_loop1_.x", function()
	--~ jdump.on()
	local sum = 0
	for i = 1, n4 do
		sum = sum + rand_vecs[i%n3].x
		sum = sum + rand_vecs[i%n3].y
		sum = sum + rand_vecs[i%n3].z
	end
	--~ jdump.off()
	print(sum)
end)

meassure("vec_index_loop2_[1]", function()
	local sum = 0
	for i = 1, n4 do
		sum = sum + rand_vecs[i%n3][1]
		sum = sum + rand_vecs[i%n3][2]
		sum = sum + rand_vecs[i%n3][3]
	end
	print(sum)
end, true)

print("benchmarking done.")

```

#### Results:

master:
```
benchmarking...
Lua: LuaJIT 2.1.0-beta3

create_rand_vecs:
time taken: 271.747 ms

vec_add_loop1_.add:
50006693.021157
time taken: 2292.325 ms

vec_add_loop1_.add:
50006693.021157
time taken: 2279.754 ms

vec_add_loop1_.add:
50006693.021157
time taken: 2300.492 ms

vec_add2_loop1_.add:
0
time taken: 2548.892 ms

vec_add2_loop1_.add:
0
time taken: 2057.055 ms

vec_index_loop1_.x:
6.8982152930833e+109
time taken: 1709.975 ms

benchmarking done.
```

this PR:
```
benchmarking...
Lua: LuaJIT 2.1.0-beta3

create_rand_vecs:
time taken: 305.866 ms

vec_add_loop1_.add:
49991247.564712
time taken: 2245.698 ms

vec_add_loop1_.add:
49991247.564712
time taken: 2298.445 ms

vec_add_loop1_.add:
49991247.564712
time taken: 2304.906 ms

vec_add2_loop1_.add:
0
time taken: 2319.513 ms

vec_add2_loop1_.add:
0
time taken: 2274.248 ms

vec_index_loop1_.x:
6.9058452003397e+109
time taken: 1663.239 ms

benchmarking done.
```
also this PR, but all benchmarks enabled:
```
benchmarking...
Lua: LuaJIT 2.1.0-beta3

create_rand_vecs:
time taken: 312.248 ms

vec_add_loop1_.add:
49964549.735267
time taken: 2248.874 ms

vec_add_loop2_:add:
49964549.735267
time taken: 1982.679 ms

vec_add_loop3_+:
49964549.735267
time taken: 1954.441 ms

vec_add_loop1_.add:
49964549.735267
time taken: 2237.673 ms

vec_add_loop1_.add:
49964549.735267
time taken: 2220.716 ms

vec_add2_loop1_.add:
0
time taken: 2404.558 ms

vec_add2_loop2_:add:
0
time taken: 2381.15 ms

vec_add2_loop3_+:
0
time taken: 2024.108 ms

vec_add2_loop1_.add:
0
time taken: 2228.035 ms

vec_index_loop1_.x:
5.2423356774124e+208
time taken: 1663.82 ms

vec_index_loop2_[1]:
5.2423356774124e+208
time taken: 3235.125 ms

benchmarking done.
```

#### evaluation:
- `vec_add_loop1_.add` is slower than the `+` and `:add` versions. I guess this is because we operate on the local sum variable. This was also the reason why I added the `add2_loop`s.
- `vec_add2_loop2_:add` seems to be slower than `vec_add2_loop3_+`. But if I first execute the `+`, then both times are equal `¯\_(ツ)_/¯`. It looks like there's no big performance difference.
- `vec_index_loop1_.x` is much faster than `vec_index_loop2_[1]`.
- Note that I've executed these benchmarks multiple times and the results were always similar. However, one might be able to see some tiny differences by doing the benchmarks much more often and averaging the results. Results might also differ for different lua/-jit versions. I am too lazy to do better benchmarks. If you really need more performance (unlikely), you should consider not using minetest's vector helpers at all

</details>

## To do

This PR is a Ready for Review.

## How to test

- Run the unittests.
- Play with mods that use the `vector.*` helpers, and look for errors.
- Look through the code.
- Do your own performance tests.