@Desour 2021-03-08
It turned out that much of the builtin code constructed vectors using `{x=...}`, which eg. caused the `pos` in `minetest.register_on_dignode`'s callback not to have a metatable. These problems should be fixed now.