@Desour 2021-03-11
Thanks for the review!

>How do I run the unit tests?

- Install busted (eg. with luarocks).
- Run `$ busted builtin` in your minetest dir.

See also: https://github.com/minetest/minetest/pull/9184#issue-351096122