@appgurueu 2021-03-17
Not a fan of the rename. If you don't want the repetition of `vector.is_vector`, cut it down to `vector.is`.