@Desour 2021-03-17
Renamed `is_real` to `is`.