@Desour 2021-04-05
Squashed and rebased.
(I had to change an expected result in the unittest for `vector.round` because of #10803.)