@Desour 2021-04-05
With the help of #10323, you can now do `tostring(v)` to turn a vector into a string.
(Thanks for merging that PR!)