@sorcerykid 2021-04-22
What about `vector.is_valid()`? The "is" by itself (at least to me, being a native English speaker) suggests a comparison, even though it's a validation. Another possibility would be "check" instead of "is_valid".

Other than that, this is shaping up to be a really nice PR!