@Desour 2021-04-24
`check` sounds nice, I've adopted that.

(Also squashed and rebased.)