@lhofhansl 2021-04-25
I like `is_valid` it's clear what it does and what it results (boolean). `check` is ambiguous (what does it check? what does it return? would it throw an exception?, etc).
Same with `is`. Nothing wrong with `'is_vector`, the meaning is clear from its name.

But I do not feel strongly, and we do not need to restart the discussion. :)

Idea is good. Lemme do a quick a review.