@appgurueu 2021-05-07
I used the following simplified code to run some benchmarks:

```lua
local socket = require("socket")
local samples = 1e7
local assert = assert
local function bench(func)
    local time = socket.gettime()
    for _ = 1, samples do
        assert(func().z == 3)
    end
    return socket.gettime() - time
end

local metatable = {z = 3}
metatable.__index = metatable

local function new()
    return setmetatable({x = 1, y = 2}, metatable)
end

local function new_v()
    local v = setmetatable({x = 1, y = 2}, metatable)
    return v
end

local setmetatable = setmetatable
local function new_local()
    return setmetatable({x = 1, y = 2}, metatable)
end

local t1, t2, t3 = bench(new), bench(new_v), bench(new_local)

print(t1/t2, t2/t3, t1/t3)
```

Results for Lua: `0.98920939595292        0.95346128018269        0.94317285703401` => Doesn't matter much. Global call appears to be fastest however.

Results for LuaJIT: `1.0741184936436 1.9369676320273 2.0805327551495` => Global call is a bit slower than `local v`, which again is a lot slower than the localized `setmetatable`, which is roughly 2x faster than the global.

Conclusion: Include `local setmetatable = setmetatable` in your code.