@Desour 2021-05-10
Thanks for your idea!

I've tried to reproduce.
(I've meassured with `chronos.nanotime()` (https://luarocks.org/modules/ldrumm/chronos) because I'm not sure where "socket" comes from.)
With luajit-2.0.5 I get no significant difference. Example output:
`1.0007221295697	1.0106367633551	1.0113665740461`
With luajit-2.1.0-beta3 tough, `local v` seems to be slower than the others:
`0.56354084118264	1.7898922873673	1.0086774052493`

So, I'll use the `local setmetatable = setmetatable`.